myapp/requirements.txt
fastapi
uvicorn[standard]
Pillow
pypiwin32
tencentcloud-sdk-python